Initial Scene Management Implementation
Intial implementation of the Scene Management pipeline and its integration into the SDL2_Engine. Modified the Testing Grounds project to make use of the new setup and tested the implementation. Initial results have a local texture loaded through the Local Resource manager and displayed to the screen using the Rendering interface being updated every cycle through the Scene Manager interface. Promising first results but needs more testing

+++ SDL2_Engine/Projects/SDL2_Engine/src/Audio/Audio.cpp
@@ -395,7 +395,7 @@ namespace SDL2_Engine {
 	*/
 	bool Audio::increaseSFXChannels() {
 		//Check if the maximum can be reached
-		if (mData->maximumSFXChannels != -1 && mData->SFXChannels >= mData->maximumSFXChannels) 
+		if (mData->maximumSFXChannels != -1 && (int)mData->SFXChannels >= mData->maximumSFXChannels) 
 			return false;
 
 		//Allocate a new channel
